* 64 bit quicktime support
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@613 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/include/avdec_private.h
@@ -596,7 +596,7 @@ bgav_input_context_t * bgav_input_open_vcd(const char * device);
 void bgav_input_close(bgav_input_context_t * ctx);
 void bgav_input_destroy(bgav_input_context_t * ctx);
 
-void bgav_input_skip(bgav_input_context_t *, int);
+void bgav_input_skip(bgav_input_context_t *, int64_t);
 
 bgav_input_context_t * bgav_input_create(bgav_options_t * opt);
